добавляет разбиение основного кода на несколько модулей
